Fixed bug when editing address. Now sending Int instead of Double. More code cleanup

+++ FireFlight/FireFlight/Models/UserAddress.swift
@@ -15,7 +15,7 @@ struct UserAddress: Codable {
     var longitude: Double
     var address: String
     var label: String?
-    var radius: Double
+    var radius: Int
     
     enum CodingKeys: String, CodingKey {
         case id
